refactor(bootstrap): A66 steps 2-4 - extract option functions into per-group files
Move With* option functions out of bootstrap.go into five dedicated files:
  options_assembly.go  — WithConfig/WithAssembly/WithAssemblyID/WithHookTimeout/WithHookObserver
  options_http.go      — WithRouterOptions/WithTracer/WithRateLimiter/WithCircuitBreaker/
                         WithSecurityHeadersOptions/WithErrorRedactor/WithHealthChecker/
                         WithReadyzDeadline/WithAdapterInfo/WithHealthRoutes
  options_events.go    — WithWorkers/WithPublisher/WithSubscriber/WithConsumerMiddleware/
                         WithEventRouterReadyTimeout/WithDisableObservabilityRestore
  options_lifecycle.go — WithLifecycle/WithLifecycleDefaultStartTimeout/
                         WithLifecycleDefaultStopTimeout/WithManagedCloser/
                         WithShutdownTimeout/WithPreShutdownDelay
  options_metrics.go   — WithMetricsProvider

bootstrap.go retains only: struct definitions, New/Lifecycle/MetricsProvider/Run,
validation helpers, and utility functions. All public signatures unchanged.
