+++ runtime/bootstrap/bootstrap.go
@@ -27,7 +27,6 @@ import (
 	"github.com/ghbvf/gocell/kernel/wrapper"
 	"github.com/ghbvf/gocell/pkg/errcode"
 	"github.com/ghbvf/gocell/runtime/config"
-	"github.com/ghbvf/gocell/runtime/http/middleware"
 	"github.com/ghbvf/gocell/runtime/http/router"
 	metricsmiddleware "github.com/ghbvf/gocell/runtime/observability/metrics"
 	"github.com/ghbvf/gocell/runtime/observability/tracing"
@@ -135,434 +134,6 @@ type Bootstrap struct {
 type namedChecker struct {
 	name string                      // unique identifier shown in /readyz?verbose output
 	fn   func(context.Context) error // nil return = healthy; non-nil = unhealthy
-}
-
-// ---------------------------------------------------------------------------
-// Option functions — assembly group
-// ---------------------------------------------------------------------------
-
-// WithConfig sets the YAML config path and environment prefix.
-func WithConfig(yamlPath, envPrefix string) Option {
-	return func(b *Bootstrap) {
-		b.assembly.configPath = yamlPath
-		b.assembly.envPrefix = envPrefix
-	}
-}
-
-// WithAssembly sets a pre-built CoreAssembly.
-func WithAssembly(asm *assembly.CoreAssembly) Option {
-	return func(b *Bootstrap) {
-		b.assembly.core = asm
-	}
-}
-
-// WithAssemblyID sets only the `cell_id` label used by the auto-wired HTTP
-// metrics collector (R2). It does not change assembly identity, cell metadata,
-// routing, health, tracing, or non-HTTP metrics.
-//
-// Recommended to set this matching asm.ID() when using WithAssembly(asm);
-// omit to reuse assembly ID (auto-derived). Explicit value overrides
-// assembly-derived.
-//
-// When neither WithAssemblyID nor WithAssembly is used, Bootstrap defaults
-// to "default" (the ID of the auto-built assembly).
-func WithAssemblyID(id string) Option {
-	return func(b *Bootstrap) {
-		b.assembly.assemblyID = id
-	}
-}
-
-// WithHookTimeout configures the per-hook deadline for the default
-// assembly built when no WithAssembly option is supplied. Zero uses
-// assembly.DefaultHookTimeout. Negative values disable per-hook
-// timeouts entirely.
-//
-// When WithAssembly is used, the pre-built assembly's Config.HookTimeout
-// takes precedence — this option has no effect. For pre-built assemblies,
-// set the value directly on assembly.Config when constructing.
-func WithHookTimeout(d time.Duration) Option {
-	return func(b *Bootstrap) {
-		b.assembly.hookTimeout = d
-		b.assembly.hookTimeoutSet = true
-	}
-}
-
-// WithHookObserver registers a cell lifecycle hook observer for the
-// default assembly built when no WithAssembly option is supplied.
-//
-// When WithAssembly is used, the pre-built assembly's Config.HookObserver
-// takes precedence — this option has no effect. For pre-built assemblies,
-// set the observer directly on assembly.Config when constructing.
-//
-// A nil observer (including a typed nil wrapping a nil concrete pointer)
-// is equivalent to not calling this option.
-func WithHookObserver(obs cell.LifecycleHookObserver) Option {
-	return func(b *Bootstrap) {
-		if cell.IsNilHookObserver(obs) {
-			return
-		}
-		b.assembly.hookObserver = obs
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Option functions — http group
-// ---------------------------------------------------------------------------
-
-// WithRouterOptions passes options to the router builder.
-func WithRouterOptions(opts ...router.Option) Option {
-	return func(b *Bootstrap) {
-		b.http.routerOpts = append(b.http.routerOpts, opts...)
-	}
-}
-
-// WithTracer enables distributed tracing. The tracer is forwarded to
-// router.WithTracer (the single HTTP request span owner) and stored on
-// Bootstrap.http.wrapperTracer so eventrouter.ContractTracingMiddleware can create
-// consumer-side wrapper.WrapConsumer spans. Without this option, HTTP tracing
-// is disabled and WrapConsumer falls back to wrapper.NoopTracer{}; a slog.Warn
-// is emitted at bootstrap time so ops notice the silent degrade.
-//
-// ref: go-zero — observability configuration at app level
-func WithTracer(t tracing.Tracer) Option {
-	return func(b *Bootstrap) {
-		b.http.routerOpts = append(b.http.routerOpts,
-			router.WithTracer(t),
-			// Skip span creation for canonical infra probe endpoints
-			// (/healthz, /readyz, /metrics) so high-rate liveness/readiness
-			// probes do not pollute trace storage. Pre-PR-A14b this was
-			// implicit because probe routes lived on the outer mux and
-			// bypassed Tracing entirely; with per-listener routers the
-			// HealthListener's full middleware chain runs, so we must
-			// install the filter explicitly here.
-			router.WithTracingOptions(middleware.WithProbeFilter(middleware.DefaultProbeFilter)),
-		)
-		b.http.wrapperTracer = t
-	}
-}
-
-// WithRateLimiter enables per-IP rate limiting for HTTP requests. The limiter
-// is forwarded to the router's middleware chain via router.WithRateLimiter.
-// If the limiter implements lifecycle.ContextCloser or io.Closer
-// (e.g. adapters/ratelimit.Limiter), Bootstrap registers it for teardown on
-// shutdown and startup rollback. ContextCloser is preferred so the shared
-// shutCtx budget flows through to the resource.
-//
-// Note: the rate limiter uses the client IP from RealIP middleware as the
-// bucket key. Ensure WithTrustedProxies is correctly configured; an overly
-// permissive trust list allows X-Forwarded-For spoofing, which bypasses
-// rate limiting.
-//
-// ref: go-zero — rate limiting configuration at app level
-// ref: uber-go/fx lifecycle OnStop(ctx) — ContextCloser preferred over io.Closer
-func WithRateLimiter(rl middleware.RateLimiter) Option {
-	return func(b *Bootstrap) {
-		b.http.routerOpts = append(b.http.routerOpts, router.WithRateLimiter(rl))
-		b.lc.closers = append(b.lc.closers, rl)
-	}
-}
-
-// WithCircuitBreaker enables circuit breaker protection for HTTP requests.
-// The breaker is forwarded to the router's middleware chain via
-// router.WithCircuitBreaker. If the breaker implements lifecycle.ContextCloser
-// or io.Closer, Bootstrap registers it for teardown on shutdown and startup
-// rollback. ContextCloser is preferred so the shared shutCtx budget flows
-// through to the resource.
-//
-// A nil cb is rejected at Run() time with a fatal error so operators are not
-// silently left without circuit-breaker protection.
-//
-// ref: go-zero — resilience middleware configuration at app level
-// ref: kubernetes/kubernetes apiserver — option fail-fast at startup
-// ref: uber-go/fx lifecycle OnStop(ctx) — ContextCloser preferred over io.Closer
-func WithCircuitBreaker(cb middleware.Allower) Option {
-	return func(b *Bootstrap) {
-		if cb == nil || middleware.IsTypedNilAllower(cb) {
-			b.http.circuitBreakerNil = true
-			return
-		}
-		b.http.routerOpts = append(b.http.routerOpts, router.WithCircuitBreaker(cb))
-		b.lc.closers = append(b.lc.closers, cb)
-	}
-}
-
-// WithSecurityHeadersOptions configures HSTS and other security header
-// directives. This is a convenience wrapper around
-// WithRouterOptions(router.WithSecurityHeadersOptions(...)).
-//
-// ref: unrolled/secure — configurable HSTS directives via struct fields
-func WithSecurityHeadersOptions(opts ...middleware.SecurityHeadersOption) Option {
-	return func(b *Bootstrap) {
-		b.http.routerOpts = append(b.http.routerOpts, router.WithSecurityHeadersOptions(opts...))
-	}
-}
-
-// WithErrorRedactor installs a wrapper.ErrorRedactor that scrubs error text
-// before it reaches span.RecordError on HTTP request spans and consumer-side
-// CONSUME spans. A nil fn disables redaction (identity semantics).
-//
-// Use when strict source-side sanitisation is required (regulated
-// environments); otherwise leave unset and let the OTel span processor /
-// exporter filter handle scrubbing at export time.
-func WithErrorRedactor(fn wrapper.ErrorRedactor) Option {
-	return func(b *Bootstrap) {
-		if fn != nil {
-			b.http.errorRedactor = fn
-			b.http.routerOpts = append(b.http.routerOpts, router.WithTracingOptions(middleware.WithErrorRedactor(fn)))
-		}
-	}
-}
-
-// WithHealthChecker registers a named readiness checker that contributes to
-// aggregate /readyz and appears in `/readyz?verbose` responses. Use this to
-// wire adapter health probes (e.g., conn.Health for RabbitMQ) without
-// bootstrap depending on adapter types.
-//
-// Accepts func(context.Context) error so callers can honour the /readyz probe
-// deadline. Validation (empty name, nil fn) is deferred to Run() where it fires
-// at Step 0 before any component starts, returning an error directly.
-func WithHealthChecker(name string, fn func(context.Context) error) Option {
-	return func(b *Bootstrap) {
-		b.http.healthCheckers = append(b.http.healthCheckers, namedChecker{name: name, fn: fn})
-	}
-}
-
-// WithReadyzDeadline overrides the per-probe deadline for /readyz. All
-// registered checkers must complete within this duration; checkers that exceed
-// it are reported as status="timeout". A zero or negative value uses the
-// health.Handler default (5 s, Kubernetes readiness probe convention).
-//
-// ref: k8s.io/apiserver/pkg/server/healthz — server-side readyz deadline
-// independent of the kubelet HTTP connection deadline.
-func WithReadyzDeadline(d time.Duration) Option {
-	return func(b *Bootstrap) {
-		b.http.readyzDeadline = d
-	}
-}
-
-// WithAdapterInfo sets static adapter configuration metadata that is exposed
-// in /readyz?verbose output. Helps operators verify which storage/bus backends
-// are active without inspecting application logs.
-func WithAdapterInfo(info map[string]string) Option {
-	return func(b *Bootstrap) {
-		b.http.adapterInfo = info
-	}
-}
-
-// WithHealthRoutes accumulates HealthRouteGroupOption values that customise
-// the framework-owned /healthz, /readyz, and /metrics route groups. The
-// canonical use cases are:
-//
-//	bootstrap.WithHealthRoutes(bootstrap.WithMetricsHandler(promHandler))
-//	bootstrap.WithHealthRoutes(bootstrap.WithReadyzVerboseToken(token))
-//	bootstrap.WithHealthRoutes(bootstrap.WithReadyzVerboseDisabled())
-//
-// Multiple WithHealthRoutes calls accumulate; later options for the same
-// concern (metrics handler, verbose-token, verbose-disabled) overwrite earlier
-// ones in the order they were appended. Pass nil-valued options at your peril —
-// they overwrite any previously-set value with the zero value.
-//
-// PR-A35 / PR269 round-3 strict semantics: a request with ?verbose= but no
-// matching readyz verbose-token / disabled flag yields 401
-// ErrReadyzVerboseDenied at the health handler layer, never a silent
-// downgrade to plain 200.
-func WithHealthRoutes(opts ...HealthRouteGroupOption) Option {
-	return func(b *Bootstrap) {
-		b.http.healthRouteGroupOpts = append(b.http.healthRouteGroupOpts, opts...)
-	}
-}
-
-// JWT authentication is wired through a typed []cell.ListenerAuth chain mounted
-// on a listener via WithListener:
-//
-//	bootstrap.WithListener(cell.PrimaryListener, addr, []cell.ListenerAuth{cell.NewAuthJWT(verifier)})
-//	bootstrap.WithListener(cell.PrimaryListener, addr, []cell.ListenerAuth{cell.NewAuthJWTFromAssembly(asm)})
-//
-// The previous bootstrap.WithAuthMiddleware / WithAuthDiscovery options have
-// been removed (PR-A14b round-3 F3): one mechanism, one source of truth. The
-// listener's auth chain carries the verifier; Bootstrap installs the router-aware
-// AuthMiddleware at phase5 so Public / PasswordResetExempt route bypass works
-// for free.
-
-// PR-A35's bootstrap-level WithVerboseDisabled has been removed (PR-A14b
-// round-3 F3 collapse): use WithHealthRoutes(WithReadyzVerboseDisabled()).
-// The single-mechanism rule applies — verbose-mode configuration lives on
-// the readyz route group exclusively.
-
-// ---------------------------------------------------------------------------
-// Option functions — events group
-// ---------------------------------------------------------------------------
-
-// WithWorkers adds background workers.
-func WithWorkers(ws ...worker.Worker) Option {
-	return func(b *Bootstrap) {
-		b.events.workers = append(b.events.workers, ws...)
-	}
-}
-
-// WithPublisher sets the outbox.Publisher used for event publishing.
-//
-// ref: uber-go/fx app.go — Option pattern; each Option targets a single concern.
-func WithPublisher(p outbox.Publisher) Option {
-	return func(b *Bootstrap) {
-		b.events.publisher = p
-	}
-}
-
-// WithSubscriber sets the outbox.Subscriber used for event consumption.
-//
-// ref: uber-go/fx app.go — Option pattern; each Option targets a single concern.
-func WithSubscriber(s outbox.Subscriber) Option {
-	return func(b *Bootstrap) {
-		b.events.subscriber = s
-	}
-}
-
-// WithConsumerMiddleware registers subscriber-side middleware applied to every
-// topic handler before it is passed to the underlying Subscriber.Subscribe call.
-// Middleware is applied in registration order; each entry wraps the next, so the
-// first registered middleware is outermost at invocation time.
-//
-// Typical use: inject ConsumerBase.AsMiddleware so every consumer inherits
-// two-phase Claimer idempotency, backoff retry, and DLX routing without each
-// slice wiring it individually. Observability context restoration
-// (entry.Observability → ctx) is the outermost step inside
-// outbox.SubscriberWithMiddleware.Subscribe, so middleware registered here
-// always sees a context populated with trace_id/request_id/correlation_id.
-//
-// ref: ThreeDotsLabs/watermill message/router.go — AddMiddleware wraps handlers
-// at router level; MassTransit UseMessageRetry — pipeline middleware at
-// receive-endpoint configuration.
-func WithConsumerMiddleware(mw ...outbox.SubscriptionMiddleware) Option {
-	return func(b *Bootstrap) {
-		b.events.consumerMiddleware = append(b.events.consumerMiddleware, mw...)
-	}
-}
-
-// WithEventRouterReadyTimeout overrides the EventRouter Phase-3 ready-wait
-// budget. A non-positive value disables the bound (router waits indefinitely
-// until ctx cancel). Default: eventrouter.DefaultReadyTimeout (30s).
-//
-// On timeout, Bootstrap.Run returns an error listing not-ready
-// "consumerGroup/topic" pairs so operators can pinpoint the stuck subscription.
-func WithEventRouterReadyTimeout(d time.Duration) Option {
-	return func(b *Bootstrap) {
-		b.events.routerReadyTimeoutSet = true
-		b.events.routerReadyTimeout = d
-	}
-}
-
-// WithDisableObservabilityRestore prevents the consumer-side
-// ObservabilityContextMiddleware from restoring request_id / correlation_id /
-// trace_id from outbox entry metadata into the handler context. The kill
-// switch for the consume-side observability bridge — set this only when
-// integrating with a custom observability stack that resets context keys
-// itself.
-func WithDisableObservabilityRestore() Option {
-	return func(b *Bootstrap) {
-		b.events.disableObservabilityRestore = true
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Option functions — lifecycle group
-// ---------------------------------------------------------------------------
-
-// WithLifecycle registers a hook-registration callback invoked during New()
-// (after all options are applied, as part of lifecycle initialisation). Use
-// for composition-root Hook registration without needing a Bootstrap reference.
-// Multiple WithLifecycle options and direct b.Lifecycle().Append() calls
-// accumulate in the order they are applied.
-func WithLifecycle(fn func(lc Lifecycle)) Option {
-	return func(b *Bootstrap) {
-		if fn != nil {
-			b.lc.lifecycleRegistrars = append(b.lc.lifecycleRegistrars, fn)
-		}
-	}
-}
-
-// WithLifecycleDefaultStartTimeout overrides the per-hook default StartTimeout.
-// Zero value retains DefaultStartTimeout (30s). Negative disables default timeout
-// (hooks without own StartTimeout will block indefinitely).
-func WithLifecycleDefaultStartTimeout(d time.Duration) Option {
-	return func(b *Bootstrap) { b.lc.lifecycleDefaultStartTimeout = d }
-}
-
-// WithLifecycleDefaultStopTimeout mirrors WithLifecycleDefaultStartTimeout for StopTimeout.
-// Zero value retains DefaultStopTimeout (10s). Negative disables default timeout.
-func WithLifecycleDefaultStopTimeout(d time.Duration) Option {
-	return func(b *Bootstrap) { b.lc.lifecycleDefaultStopTimeout = d }
-}
-
-// WithManagedCloser registers an adapter or resource that implements
-// lifecycle.ContextCloser for LIFO teardown during graceful shutdown.
-// The shared shutCtx budget propagates directly to c.Close(ctx), so the
-// resource participates in the same shutdown deadline as all other components.
-//
-// Use this instead of a bare defer c.Close() so that:
-//
-//   - The resource is closed in LIFO order after HTTP and worker shutdown.
-//   - The shared shutdownTimeout ctx is honoured (not an arbitrary timeout).
-//   - Startup rollback also triggers the teardown on phase failures.
-//
-// A nil c is silently ignored (consistent with addCloser semantics).
-//
-// ref: uber-go/fx Lifecycle.Append OnStop(ctx) — managed teardown registration.
-// ref: sigs.k8s.io/controller-runtime pkg/manager/internal.go engageStopProcedure LIFO.
-func WithManagedCloser(c kernellifecycle.ContextCloser) Option {
-	return func(b *Bootstrap) {
-		if c == nil {
-			return
-		}
-		b.lc.closers = append(b.lc.closers, c)
-	}
-}
-
-// WithShutdownTimeout overrides the default graceful shutdown timeout.
-func WithShutdownTimeout(d time.Duration) Option {
-	return func(b *Bootstrap) {
-		b.lc.shutdownTimeout = d
-	}
-}
-
-// WithPreShutdownDelay sets a delay between marking /readyz as 503 and
-// starting the HTTP server shutdown. This gives load balancers (e.g.,
-// Kubernetes kube-proxy) time to observe the unhealthy readiness probe
-// and stop routing new traffic before the server closes connections.
-//
-// Default is 0 (no delay). Typical Kubernetes deployments use 3-5 seconds.
-// The delay counts toward the total shutdownTimeout budget (not additive).
-//
-// ref: Kubernetes pod shutdown — preStop counts toward terminationGracePeriodSeconds
-func WithPreShutdownDelay(d time.Duration) Option {
-	return func(b *Bootstrap) {
-		b.lc.preShutdownDelay = d
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Option functions — metrics group
-// ---------------------------------------------------------------------------
-
-// WithMetricsProvider registers a provider-neutral metrics backend used by
-// components that need to emit counters/histograms through a common
-// abstraction (hook dispatcher drop counters, OTel pool-stats collector,
-// custom caller-registered metrics). Pass nil or omit the option to use
-// kernel/observability/metrics.NopProvider (no emission).
-//
-// Callers can read b.MetricsProvider() to register additional metrics
-// against the same backend — useful when cmd/* builds both an HTTP
-// Collector and a relay Collector on the same Provider instance.
-//
-// ref: opentelemetry-go otel.GetMeterProvider@main — single global
-// provider entry point; GoCell exposes it per-Bootstrap instance to avoid
-// mutable global state.
-func WithMetricsProvider(p kernelmetrics.Provider) Option {
-	return func(b *Bootstrap) {
-		if p != nil {
-			b.metrics.provider = p
-		}
-	}
 }
 
 // ---------------------------------------------------------------------------
